backend/requirements.txt
# FastAPI and ASGI server
fastapi[all]>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0

# OpenAI Agents SDK (for agent-based LLM responses)
openai-agents>=0.2.9,<1.0.0

# Sentence Transformers for local embeddings (BAAI/bge-base-en-v1.5, 768 dims)
sentence-transformers>=2.7.0,<4.0.0
torch>=2.5.0,<3.0.0

# Qdrant vector database client
qdrant-client>=1.14.1,<2.0.0

# SQLAlchemy ORM and PostgreSQL driver
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.1,<2.0.0

# Pydantic for validation (updated for openai-agents compatibility)
pydantic>=2.10,<3.0
pydantic-settings>=2.5.2

# Rate limiting
slowapi>=0.1.9,<1.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Markdown parsing and text splitting
markdown-it-py>=3.0.0,<4.0.0
langchain-text-splitters>=0.0.1,<1.0.0

# YAML parsing for frontmatter
pyyaml>=6.0.1,<7.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0
httpx>=0.27.1  # For async testing (updated for mcp compatibility)

# Error tracking (optional)
sentry-sdk[fastapi]>=2.0.0,<3.0.0
